Add files via upload
Code Functionality: This code is designed to read the RFID card's code. If the card's code matches the saved one in the ESP8266, the servo motor will rotate 180° and return to its original position after 5 seconds.

Pin Mapping: Note that pin 4 corresponds to D2 pin on the ESP8266 D1 Mini as wired in the diagram.

Before Usage:

UID Code Modification: Before using the code, it's essential to change the UID code in the "DOOR_OPENER_CODE_ESP8266_SERVO.ino" file, specifically on line 14. This UID code should match the one obtained from the serial monitor during the scanning of cards with the "DumpInfoESP8266.ino" code.
UID Code Format: The UID code will be provided in this format: "54 D9 82 24". It needs to be modified to this format: "0x54, 0xD9, 0x82, 0x24" before replacing it in the code.